+++ Editor/Processors/BaseProcessor.cs
@@ -10,6 +10,8 @@ namespace Hertzole.CecilAttributes.Editor
 
         public abstract bool AllowEditor { get; }
 
+        public virtual bool EditorOnly { get { return false; } }
+
         public abstract bool IsValidClass(TypeDefinition type);
 
         public abstract (bool success, bool dirty) ProcessClass(ModuleDefinition module, TypeDefinition type);